requirements-prod.txt
# 生产环境依赖
-r requirements.txt

# 生产服务器
gunicorn>=21.2.0
uvicorn[standard]>=0.23.0

# 监控和日志
prometheus-client>=0.17.0
structlog>=23.1.0
sentry-sdk[fastapi]>=1.32.0

# 性能优化
orjson>=3.9.0
ujson>=5.8.0

# 缓存
redis[hiredis]>=4.6.0

# 数据库优化
psycopg2-binary>=2.9.7
SQLAlchemy[asyncio]>=2.0.0

# 安全
cryptography>=41.0.0
passlib[bcrypt]>=1.7.4

# 分布式追踪
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-jaeger>=1.20.0
opentelemetry-exporter-prometheus>=1.12.0rc1
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
opentelemetry-instrumentation-sqlalchemy>=0.41b0
opentelemetry-instrumentation-redis>=0.41b0

# 云存储
boto3>=1.28.0
minio>=7.1.0

# 任务队列
celery[redis]>=5.3.0

# 配置管理
pydantic-settings>=2.0.0 